requirements.txt
python-binance
pandas-stubs>=1.5.0
ta
matplotlib
requests
aiohttp>=3.9.0
websockets>=12.0
plotly>=5.18.0
python-dotenv
feedparser
tenacity
textblob

# Required for TextBlob NLP
nltk
